w3c-reports/12092014/efl/runned-tests/XMLHttpRequest/send-content-type-string-expected.txt

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

  

PASS XMLHttpRequest: send() - Content-Type 
FAIL XMLHttpRequest: send() - Content-Type 1 assert_equals: expected "application/xml;charset=windows-1252" but got "NO"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/XMLHttpRequest/send-content-type-string.htm:17:24
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
request@http://localhost:8800/XMLHttpRequest/send-content-type-string.htm:13:13
init@http://localhost:8800/XMLHttpRequest/send-content-type-string.htm:21:34
onload@http://localhost:8800/XMLHttpRequest/send-content-type-string.htm:23:5)
FAIL XMLHttpRequest: send() - Content-Type 2 assert_equals: expected "text/html;charset=windows-1252" but got "NO"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/XMLHttpRequest/send-content-type-string.htm:17:24
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
request@http://localhost:8800/XMLHttpRequest/send-content-type-string.htm:13:13
init@http://localhost:8800/XMLHttpRequest/send-content-type-string.htm:21:34
onload@http://localhost:8800/XMLHttpRequest/send-content-type-string.htm:24:5)
